+++ simpleadb/__init__.py
@@ -6,9 +6,16 @@
 # SPDX-License-Identifier: GPL-3.0-only
 #
 
-""" simpleadb init module """
+""" simpleadb init module. """
+
+from .adbprocess import AdbCommandError
+from .adbprocess import AdbCommandTimeoutExpired
 from .adbdevice import AdbDevice
-from .adbdevice import AdbCommandError
 from .adbserver import AdbServer
 
-__all__ = ['AdbDevice', 'AdbCommandError', 'AdbServer']
+__all__ = [
+    'AdbCommandError',
+    'AdbCommandTimeoutExpired',
+    'AdbDevice',
+    'AdbServer'
+]